Package Reference

Macro.get_value — Method

get_value(dict::AbstractDict, keys::Vector{Symbol})

Get the value from a dictionary based on a sequence of keys.

Arguments

dict::AbstractDict: The dictionary from which to retrieve the value.

keys::Vector{Symbol}: The sequence of keys to traverse the dictionary.

Returns

The value retrieved from the dictionary based on the given keys.

Examples

dict = Dict(:a => Dict(:b => 1, :c => Dict(:b => 2)))
get_value(dict, [:a, :b]) # returns 1
get_value(dict, [:a, :c, :b]) # returns 2

Macro.get_value_and_keys — Function

get_value_and_keys(dict::AbstractDict, target_key::Symbol, keys=Symbol[])

Recursively searches for a target key in a dictionary and returns a list of tuples containing the value associated with the target key and the keys leading to it. This function is used to replace the path to a timeseries file with the actual vector of data.

Arguments

dict::AbstractDict: The (nested) dictionary to search in.

target_key::Symbol: The key to search for.

keys=Symbol[]: (optional) The keys leading to the current dictionary.

Returns

value_keys: A list of tuples, where each tuple contains - the value associated with the target key - the keys leading to it in the nested dictionary.

Examples

dict = Dict(:a => Dict(:b => 1, :c => Dict(:b => 2)))
get_value_and_keys(dict, :b) # returns [(1, [:a, :b]), (2, [:a, :c, :b])]

Where the first element of the tuple is the value of the key :b and the second element is the list of keys to reach that value.

Macro.make — Method

make(::Type{<:VRE}, data::AbstractDict{Symbol, Any}, system::System) -> VRE

VRE is an alias for Union{SolarPV, WindTurbine}

Necessary data fields:
 - transforms: Dict{Symbol, Any}
    - id: String
    - timedata: String
- edges: Dict{Symbol, Any}
    - edge: Dict{Symbol, Any}
        - id: String
        - end_vertex: String
        - unidirectional: Bool
        - has_planning_variables: Bool
        - can_retire: Bool
        - can_expand: Bool
        - constraints: Vector{AbstractTypeConstraint}

Macro.make — Method

make(::Type{Battery}, data::AbstractDict{Symbol, Any}, system::System) -> Battery

Necessary data fields:
 - storage: Dict{Symbol, Any}
    - id: String
    - commodity: String
    - can_retire: Bool
    - can_expand: Bool
    - existing_capacity_storage: Float64
    - investment_cost_storage: Float64
    - fixed_om_cost_storage: Float64
    - storage_loss_fraction: Float64
    - min_duration: Float64
    - max_duration: Float64
    - min_storage_level: Float64
    - min_capacity_storage: Float64
    - max_capacity_storage: Float64
    - constraints: Vector{AbstractTypeConstraint}
 - edges: Dict{Symbol, Any}
    - charge_edge: Dict{Symbol, Any}
        - id: String
        - start_vertex: String
        - unidirectional: Bool
        - has_planning_variables: Bool
        - efficiency: Float64
    - discharge_edge: Dict{Symbol, Any}
        - id: String
        - end_vertex: String
        - unidirectional: Bool
        - has_planning_variables: Bool
        - can_retire: Bool
        - can_expand: Bool
        - efficiency
        - constraints: Vector{AbstractTypeConstraint}

Macro.make — Method

make(::Type{Electrolyzer}, data::AbstractDict{Symbol, Any}, system::System) -> Electrolyzer

Necessary data fields:
 - transforms: Dict{Symbol, Any}
    - id: String
    - timedata: String
    - efficiency_rate: Float64
    - constraints: Vector{AbstractTypeConstraint}
- edges: Dict{Symbol, Any}
    - h2_edge: Dict{Symbol, Any}
        - id: String
        - end_vertex: String
        - unidirectional: Bool
        - has_planning_variables: Bool
        - can_retire: Bool
        - can_expand: Bool
        - constraints: Vector{AbstractTypeConstraint}
    - e_edge: Dict{Symbol, Any}
        - id: String
        - start_vertex: String
        - unidirectional: Bool
        - has_planning_variables: Bool
        - can_retire: Bool
        - can_expand: Bool
        - constraints: Vector{AbstractTypeConstraint}

Macro.make — Method

make(::Type{FuelCell}, data::AbstractDict{Symbol, Any}, system::System) -> FuelCell

Necessary data fields:
 - transforms: Dict{Symbol, Any}
    - id: String
    - timedata: String
    - efficiency_rate: Float64
    - constraints: Vector{AbstractTypeConstraint}
- edges: Dict{Symbol, Any}
    - h2_edge: Dict{Symbol, Any}
        - id: String
        - end_vertex: String
        - unidirectional: Bool
        - has_planning_variables: Bool
        - can_retire: Bool
        - can_expand: Bool
        - constraints: Vector{AbstractTypeConstraint}
    - e_edge: Dict{Symbol, Any}
        - id: String
        - start_vertex: String
        - unidirectional: Bool
        - has_planning_variables: Bool
        - can_retire: Bool
        - can_expand: Bool
        - constraints: Vector{AbstractTypeConstraint}

Macro.make — Method

make(::Type{NaturalGasHydrogen}, data::AbstractDict{Symbol, Any}, system::System) -> NaturalGasHydrogen

Necessary data fields:
 - transforms: Dict{Symbol, Any}
    - id: String
    - timedata: String
    - efficiency_rate: Float64
    - emission_rate: Float64
    - constraints: Vector{AbstractTypeConstraint}
- edges: Dict{Symbol, Any}
    - h2_edge: Dict{Symbol, Any}
        - id: String
        - end_vertex: String
        - unidirectional: Bool
        - has_planning_variables: Bool
        - can_retire: Bool
        - can_expand: Bool
        - min_up_time: Int
        - min_down_time: Int
        - startup_cost: Float64
        - startup_fuel: Float64
        - startup_fuel_balance_id: Symbol
        - constraints: Vector{AbstractTypeConstraint}
    - ng_edge: Dict{Symbol, Any}
        - id: String
        - start_vertex: String
        - unidirectional: Bool
        - has_planning_variables: Bool
        - can_retire: Bool
        - can_expand: Bool
        - constraints: Vector{AbstractTypeConstraint}
    - co2_edge: Dict{Symbol, Any}
        - id: String
        - end_vertex: String
        - unidirectional: Bool
        - has_planning_variables: Bool
        - can_retire: Bool
        - can_expand: Bool
        - constraints: Vector{AbstractTypeConstraint}

Macro.make — Method

make(::Type{NaturalGasPower}, data::AbstractDict{Symbol, Any}, system::System) -> NaturalGasPower

Necessary data fields:
 - transforms: Dict{Symbol, Any}
    - id: String
    - timedata: String
    - heat_rate: Float64
    - emission_rate: Float64
    - constraints: Vector{AbstractTypeConstraint}
- edges: Dict{Symbol, Any}
    - e_edge: Dict{Symbol, Any}
        - id: String
        - end_vertex: String
        - unidirectional: Bool
        - has_planning_variables: Bool
        - can_retire: Bool
        - can_expand: Bool
        - min_up_time: Int
        - min_down_time: Int
        - startup_cost: Float64
        - startup_fuel: Float64
        - startup_fuel_balance_id: Symbol
        - constraints: Vector{AbstractTypeConstraint}
    - ng_edge: Dict{Symbol, Any}
        - id: String
        - start_vertex: String
        - unidirectional: Bool
        - has_planning_variables: Bool
        - can_retire: Bool
        - can_expand: Bool
        - constraints: Vector{AbstractTypeConstraint}
    - co2_edge: Dict{Symbol, Any}
        - id: String
        - end_vertex: String
        - unidirectional: Bool
        - has_planning_variables: Bool
        - can_retire: Bool
        - can_expand: Bool
        - constraints: Vector{AbstractTypeConstraint}

Macro.set_value — Method

set_value(dict::AbstractDict, keys::Vector{Symbol}, new_value)

Set the value of a nested dictionary given a list of keys.

Arguments

dict::AbstractDict: The dictionary to modify.

keys::Vector{Symbol}: A list of keys representing the path to the value to

be modified.

new_value: The new value to set.

Examples

dict = Dict(:a => Dict(:b => 1, :c => Dict(:b => 2)))
set_value(dict, [:a, :b], 3)
get_value(dict, [:a, :b]) # returns 3

Macro.struct_info — Method

struct_info(t::Type{T}) where T

Return a vector of tuples with the field names and types of a struct.

Macro.timestepbefore — Method

timestepbefore(t::Int, h::Int,subperiods::Vector{StepRange{Int64,Int64})

Determines the time step that is h steps bßefore index t in subperiod p with circular indexing.