Utilities

MacroEnergy.get_value_and_keys — Function

get_value_and_keys(dict::AbstractDict, target_key::Symbol, keys=Symbol[])

Recursively searches for a target key in a dictionary and returns a list of tuples containing the value associated with the target key and the keys leading to it. This function is used to replace the path to a timeseries file with the actual vector of data.

Arguments

dict::AbstractDict: The (nested) dictionary to search in.

target_key::Symbol: The key to search for.

keys=Symbol[]: (optional) The keys leading to the current dictionary.

Returns

value_keys: A list of tuples, where each tuple contains - the value associated with the target key - the keys leading to it in the nested dictionary.

Examples

dict = Dict(:a => Dict(:b => 1, :c => Dict(:b => 2)))
get_value_and_keys(dict, :b) # returns [(1, [:a, :b]), (2, [:a, :c, :b])]

Where the first element of the tuple is the value of the key :b and the second element is the list of keys to reach that value.

MacroEnergy.set_value — Function

set_value(dict::AbstractDict, keys::Vector{Symbol}, new_value)

Set the value of a nested dictionary given a list of keys.

Arguments

dict::AbstractDict: The dictionary to modify.

keys::Vector{Symbol}: A list of keys representing the path to the value to

be modified.

new_value: The new value to set.

Examples

dict = Dict(:a => Dict(:b => 1, :c => Dict(:b => 2)))
set_value(dict, [:a, :b], 3)
get_value(dict, [:a, :b]) # returns 3

MacroEnergy.timestepbefore — Function

timestepbefore(t::Int, h::Int,subperiods::Vector{StepRange{Int64,Int64})

Determines the time step that is h steps before index t in subperiod p with circular indexing.

MacroEnergy.struct_info — Function

struct_info(t::Type{T}) where T

Return a vector of tuples with the field names and types of a struct.

MacroEnergy.get_value — Function

get_value(dict::AbstractDict, keys::Vector{Symbol})

Get the value from a dictionary based on a sequence of keys.

Arguments

dict::AbstractDict: The dictionary from which to retrieve the value.

keys::Vector{Symbol}: The sequence of keys to traverse the dictionary.

Returns

The value retrieved from the dictionary based on the given keys.

Examples

dict = Dict(:a => Dict(:b => 1, :c => Dict(:b => 2)))
get_value(dict, [:a, :b]) # returns 1
get_value(dict, [:a, :c, :b]) # returns 2